[list_backport_pr] Add opt to open links in browser
A simple convenience method for opening all of the existing pull
requests in the browser.

Currently only works for OSX, as this makes use of the `open` shell
command, and I haven't setup this for other operating systems.  We could
use something like:

  https://github.com/copiousfreetime/launchy

But I wasn't going to go that far for something that might not be useful
to anyone else.


(transferred from ManageIQ/manageiq-release@9472133e13a2d03e9d50e01d01ba8b8d8907e327)