Use ls-remote to check for remote branches
Previously, we depended on git checkout to fail with a specific
exception and error message, which could change.

It's easier to just check the remote for the branch, such as
origin/hammer and skip that repository if it doesn't have the release
branch.

It now looks like this:

=========================== ManageIQ/container-httpd
===========================
+++ blasting ManageIQ/container-httpd...
!!! Skipping ManageIQ/container-httpd: 'origin/hammer' not found

========================= ManageIQ/container-memcached
=========================
+++ blasting ManageIQ/container-memcached...
!!! Skipping ManageIQ/container-memcached: 'origin/hammer' not found


(transferred from ManageIQ/manageiq-release@e577c6b531dabd591d12309b09e262fbe0d2b6d6)