Fix mirror of non-default branches
If a branch_mirror_override specified a branch that was not in the set
of default branches, then it wouldn't be mirrored at all. This change
now honors the expected set of mirrored branches by including the
overrides in the check.


(transferred from ManageIQ/manageiq-release@725b1949533422888b59c832cd8b8c740290ba73)