Pin json gem to 2.9.1 since 2.10.0+ is broken
In https://github.com/ruby/json/issues/752, we can see that due to a bug
in rubygems, the wrong json extension is being loaded when used in
comibination with the json_pure gem, since they share some pathing
within their gems. json 2.10.0 introduced a new, rewritten, C-extension
that is introducing this bug.