Allow blank overrides on the command line
If the caller sets a default repo-set, then the user passes --repo, then
repo will be filtered from the repo-set, however that's not always what
we want. So, this commit does two things.

1. Change the wording about repo overriding repo-set, since it's
   actually more of a filter.
2. Allow blank values for repo-set (and repo) on the command line, to
   allow overriding the default repo-set.

The second point allows --repo to then act as an actual override when
there is a default present, by allowing the removal of the repo-set.