[pull_request_blaster_outer] Prevent git show paging when using --force
--force is meant for avoiding any user input on each repo and allowing
it to proceed to create the PR. However, when showing the diff, if git
is defined to use a pager it will (and by default git uses a pager).
This blocks and waits for the user to escape the pager, which kills the
purpose of --force.

Unfortunately, minigit doesn't allow params to the git command (such as
--no-pager), and those params go to the command instead. So, this commit
introduces a raw command on the git helper which is a thin wrapper
around the `git` cli. Since the code was borrowed from the clone
command, this commit also changes the clone command to reuse it.